Move registration to new module
