+++ saleor/checkout/views/__init__.py
@@ -12,7 +12,7 @@ from .validators import (
     validate_shipping_method, validate_is_shipping_required)
 from ..core import load_checkout
 from ..forms import ShippingMethodForm
-from ...userprofile.registration.forms import LoginForm
+from ...registration.forms import LoginForm
 
 
 @load_checkout
