+++ saleor/order/views.py
@@ -14,8 +14,8 @@ from .forms import PaymentDeleteForm, PaymentMethodsForm, PasswordForm
 from .models import Order, Payment
 from .utils import check_order_status, attach_order_to_user
 from ..core.utils import get_client_ip
+from ..registration.forms import LoginForm
 from ..userprofile.models import User
-from ..userprofile.registration.forms import LoginForm
 from . import OrderStatus
 
 logger = logging.getLogger(__name__)
