+++ saleor/urls.py
@@ -4,21 +4,19 @@ from django.conf.urls.static import static
 from django.contrib.sitemaps.views import sitemap
 from django.contrib.staticfiles.views import serve
 from django.views.i18n import javascript_catalog
-
 from graphene_django.views import GraphQLView
 
 from .cart.urls import urlpatterns as cart_urls
 from .checkout.urls import urlpatterns as checkout_urls
 from .core.sitemaps import sitemaps
 from .core.urls import urlpatterns as core_urls
+from .dashboard.urls import urlpatterns as dashboard_urls
+from .data_feeds.urls import urlpatterns as feed_urls
 from .order.urls import urlpatterns as order_urls
 from .product.urls import urlpatterns as product_urls
+from .registration.urls import urlpatterns as registration_urls
 from .search.urls import urlpatterns as search_urls
-from .userprofile.registration.urls import urlpatterns as registration_urls
 from .userprofile.urls import urlpatterns as userprofile_urls
-from .data_feeds.urls import urlpatterns as feed_urls
-from .dashboard.urls import urlpatterns as dashboard_urls
-
 
 urlpatterns = [
     url(r'^', include(core_urls)),
