+++ tests/test_registration.py
@@ -1,10 +1,9 @@
-from django.shortcuts import reverse
 from django.contrib.auth import get_user_model
 from django.contrib.auth.models import AnonymousUser
+from django.core.urlresolvers import reverse
 
-from saleor.userprofile.registration.forms import LoginForm, SignupForm
+from saleor.registration.forms import LoginForm, SignupForm
 from .utils import get_redirect_location
-
 
 User = get_user_model()
